src/requirements.txt
alibabacloud_alidns20150109==3.0.1
PyYAML==6.0
apscheduler==3.7.0
requests==2.28.1
urllib3==1.26.12